PG/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose
alembic
pytest
httpx
ruff
